requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
celery==4.2.1
CairoSVG==2.1.3
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
cssselect2==0.2.1
django
django-simple-captcha
defusedxml==0.5.0
django-crontab==0.7.1
django-haystack==2.8.1
django-pagedown
html5lib==1.0.1
idna==2.6
jieba==0.39
mistune
markdown2
pdfrw==0.4
Pillow==5.1.0
Pygments
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyMySQL
pyparsing==2.2.0
Pyphen==0.9.4
pytz==2018.3
python-memcached
requests==2.18.4
selenium==3.11.0
six==1.11.0
tinycss2==0.6.1
urllib3==1.22
WeasyPrint==0.40
webencodings==0.5.1
Whoosh==2.7.4
wkhtmltopdf==0.2
django-werkzeug-debugger-runserver==0.3.1
pyOpenSSL
python-docx
scrapy
scrapy_splash
django-crispy-forms>=1.6.0
django-reversion>=2.0.0
django-formtools==2.1
future==0.15.2
httplib2==0.9.2
django-import-export==0.6.1
markdown
django-filter==2.0.0
django-simple-serializer==2.0.7
coreapi==2.3.3
django-cors-headers==2.4.0
cairocffi
uwsgi
scrapy_djangoitem